added animations with transitions to pufferfish, added collission detection for fin slap with pufferfish + detection if player gets to close for transitiona nimation, first tests with dynamic background scrolling, added maincamera scroll , extended world border

+++ src/app/scenes/gamescene.ts
@@ -6,20 +6,22 @@ import { GlobalstateserviceService } from '../services/globalstate.service';
 import { Pufferfish } from '../classes/pufferfish.class';
 
 export class Gamescene extends Phaser.Scene {
+
   private background!: Background;
   private player!: Player;
-  //private enemy_0!: Jellyfish;
-  private enemies: any[] = [];
+  private enemies: Jellyfish[] | Pufferfish[] = [];
 
   constructor(public globalStateService: GlobalstateserviceService) {
     super({ key: 'Gamescene' });
     this.background = new Background(this);
     this.player = new Player(this, globalStateService);
     this.enemies = [
-      new Jellyfish(this),
-      new Jellyfish(this),
-      new Jellyfish(this),
-      new Pufferfish(this),
+      new Jellyfish(this, globalStateService),
+      new Jellyfish(this, globalStateService),
+      new Jellyfish(this, globalStateService),
+      new Pufferfish(this, globalStateService),
+      new Pufferfish(this, globalStateService),
+      new Pufferfish(this, globalStateService),
     ];
   }
 
@@ -50,28 +52,78 @@ export class Gamescene extends Phaser.Scene {
   }
 
   create() {
+    this.physics.world.setBounds(0, 0, 1920 * 2, 1080);
     this.background.create();
-    //this.enemy_0.create()
     this.createEnemies();
     this.player.create();
+    this.cameras.main.startFollow(this.player.playerSprite, true, 0.1, 0.1);
+    this.cameras.main.setBounds(0,0, 1920 * 2, 1080);
   }
 
   override update() {
     this.player.update();
-    //this.enemy_0.update()
     this.updateEnemies();
     this.checkCollisions();
+    this.garbageCollection();
   }
 
   checkCollisions() {
-    this.enemies.forEach((enemy: Jellyfish) => {
+    this.enemies.forEach((enemy: Jellyfish | Pufferfish) => {
       if (this.physics.overlap(this.player.playerSprite, enemy.enemySprite)) {
         console.log('collision');
+      }
+      if (enemy instanceof Pufferfish) {
+        this.checkPufferfishProximity(enemy) 
+        this.checkSlapCollision(enemy)
       }
     });
     //this.poisonedBubbleCollisionCheck();
     this.whiteBubbleCollisionCheck();
   }
+
+  checkPufferfishProximity(enemy: Pufferfish) {
+    if (enemy.isAggro)
+      return;
+
+    const distance = Phaser.Math.Distance.Between(
+      this.player.playerSprite.x,
+      this.player.playerSprite.y,
+      enemy.enemySprite.x,
+      enemy.enemySprite.y
+    )
+    if (distance < 400) {
+      enemy.isAggro = true;
+      enemy.checkAggroState();
+    }
+  }
+
+  checkSlapCollision(enemy: Pufferfish) {
+    if (enemy.isDead) 
+      return;
+  
+    const distance = Phaser.Math.Distance.Between(
+      this.player.playerSprite.x,
+      this.player.playerSprite.y,
+      enemy.enemySprite.x,
+      enemy.enemySprite.y
+    )
+    if (distance < 320 && this.globalStateService.hasSlapped()) {
+      enemy.isDead = true;
+      enemy.enemySprite.setVelocityX(0)
+      enemy.checkDeathState();
+    } else {
+      this.globalStateService.hasSlapped.set(false);
+    }
+  }
+
+  garbageCollection() {
+    this.enemies.forEach((enemy) => {
+      if (enemy.hasDied) {
+        enemy.enemySprite.destroy();
+      }
+    });
+  }
+  
 
   poisonedBubbleCollisionCheck() {
     const bubbles = this.player.getPBubbles();
@@ -91,7 +143,7 @@ export class Gamescene extends Phaser.Scene {
     const bubbles = this.player.getWBubbles();
     bubbles.forEach((bubble) => {
       this.enemies.forEach((enemy: Jellyfish) => {
-        if (this.physics.overlap(bubble, enemy.enemySprite)) {
+        if (enemy instanceof Jellyfish && this.physics.overlap(bubble, enemy.enemySprite)) {
           this.globalStateService.removeWBubble(bubble);
           bubble.destroy();
           enemy.isDead = true;
